+++ CHANGELOG.md
@@ -6,13 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Added
-
 - Reload database schema on SIGHUP - @begriffs
 - Support "-" in column names - @ruslantalpa
+- Support column/node renaming `alias:column` - @ruslantalpa
 
 ### Fixed
-- Set transaction mode to READ when possible to support connecting to read replicas - @ruslantalpa
-
 - Omit Content-Type header for empty body - @begriffs
 - Prevent role from being changed twice - @begriffs
 - Use read-only transaction for read requests - @ruslantalpa